camera functionality
git-svn-id: https://politbuero.dyndns.org/svn/repos/promotion@1136 f4953187-ecf7-0310-90f3-8d0fedb0aae5

+++ koptercontrol/laserscanner.cpp
@@ -122,7 +122,7 @@ void LaserScanner::slotScanFinished(const quint32 &timestamp)
                             mScannerPoseBefore,
                             mScannerPoseAfter,
                             mScannerPoseLast,
-                            0.125 + (0.75 * i / mScanDistancesCurrent->size())
+                            (float)(0.125 + (0.75 * i / mScanDistancesCurrent->size()))
                             );
 
                 // position of the point relative to the physical laserscanner
